# qontract_api/openapi.json
{
  "components": {
    "schemas": {
      "AWSAccountManagerCreateAccountRequest": {
        "description": "Request to create a plain AWS account.\n\nHandles only AWS-level operations: create org account, describe,\ntag, set alias. One payer account + one account request per call.",
        "properties": {
          "account_request": {
            "$ref": "#/components/schemas/AWSAccountRequest",
            "description": "Account to create"
          },
          "default_tags": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Default tags for the new account",
            "title": "Default Tags",
            "type": "object"
          },
          "dry_run": {
            "default": true,
            "description": "If True, only calculate actions without executing.",
            "title": "Dry Run",
            "type": "boolean"
          },
          "organization_account_role": {
            "default": "OrganizationAccountAccessRole",
            "description": "Role to assume in the new account",
            "title": "Organization Account Role",
            "type": "string"
          },
          "payer_account": {
            "$ref": "#/components/schemas/AWSPayerAccount",
            "description": "Payer account with credentials"
          }
        },
        "required": [
          "payer_account",
          "account_request"
        ],
        "title": "AWSAccountManagerCreateAccountRequest",
        "type": "object"
      },
      "AWSAccountManagerCreateIAMUserRequest": {
        "description": "Request to create an IAM user in an AWS account.\n\nAssumes into the account via the payer's manager role and creates\nan IAM user with the specified policy. Credentials are saved to Vault.",
        "properties": {
          "account_name": {
            "description": "Name of the target account",
            "title": "Account Name",
            "type": "string"
          },
          "account_uid": {
            "description": "AWS account ID of the target account",
            "title": "Account Uid",
            "type": "string"
          },
          "dry_run": {
            "default": true,
            "description": "If True, only calculate actions without executing.",
            "title": "Dry Run",
            "type": "boolean"
          },
          "organization_account_role": {
            "default": "OrganizationAccountAccessRole",
            "description": "Role to assume in the target account",
            "title": "Organization Account Role",
            "type": "string"
          },
          "payer_account": {
            "$ref": "#/components/schemas/AWSPayerAccount",
            "description": "Payer account with credentials"
          },
          "policy_arn": {
            "description": "Policy ARN to attach to the IAM user",
            "title": "Policy Arn",
            "type": "string"
          },
          "secret_vault_path": {
            "description": "Vault path for storing IAM credentials",
            "title": "Secret Vault Path",
            "type": "string"
          },
          "user_name": {
            "description": "Name of the IAM user to create",
            "title": "User Name",
            "type": "string"
          }
        },
        "required": [
          "payer_account",
          "account_name",
          "account_uid",
          "user_name",
          "policy_arn",
          "secret_vault_path"
        ],
        "title": "AWSAccountManagerCreateIAMUserRequest",
        "type": "object"
      },
      "AWSAccountManagerReconcileRequest": {
        "description": "Request to reconcile a single AWS account.\n\nIf ``organization`` is present, the account is treated as an organization\naccount (payer credentials + role assumption). Otherwise it is a standalone\naccount with direct credentials.",
        "properties": {
          "account_name": {
            "description": "Account name",
            "title": "Account Name",
            "type": "string"
          },
          "alias": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Desired account alias",
            "title": "Alias"
          },
          "automation_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payer's manager role ARN (required for org accounts)",
            "title": "Automation Role"
          },
          "automation_token": {
            "$ref": "#/components/schemas/Secret",
            "description": "Credentials (payer's for org accounts, own for non-org)"
          },
          "default_tags": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Default tags for the account",
            "title": "Default Tags",
            "type": "object"
          },
          "dry_run": {
            "default": true,
            "description": "If True, only calculate actions without executing.",
            "title": "Dry Run",
            "type": "boolean"
          },
          "enterprise_support": {
            "default": false,
            "description": "Whether enterprise support is required",
            "title": "Enterprise Support",
            "type": "boolean"
          },
          "organization": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AWSAccountOrganization"
              },
              {
                "type": "null"
              }
            ],
            "description": "Organization details (OU + tags). If set, account is org-managed."
          },
          "organization_account_role": {
            "default": "OrganizationAccountAccessRole",
            "description": "Role to assume in org account",
            "title": "Organization Account Role",
            "type": "string"
          },
          "payer_uid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Payer account UID (required for org accounts \u2014 role assumed on payer, not org account)",
            "title": "Payer Uid"
          },
          "quotas": {
            "description": "Desired service quotas",
            "items": {
              "$ref": "#/components/schemas/AWSQuota"
            },
            "title": "Quotas",
            "type": "array"
          },
          "resources_default_region": {
            "description": "Default AWS region",
            "title": "Resources Default Region",
            "type": "string"
          },
          "security_contact": {
            "$ref": "#/components/schemas/AWSSecurityContact",
            "description": "Desired security contact"
          },
          "supported_deployment_regions": {
            "description": "Desired enabled regions",
            "items": {
              "type": "string"
            },
            "title": "Supported Deployment Regions",
            "type": "array"
          },
          "uid": {
            "description": "AWS account ID",
            "title": "Uid",
            "type": "string"
          }
        },
        "required": [
          "account_name",
          "uid",
          "automation_token",
          "resources_default_region",
          "security_contact"
        ],
        "title": "AWSAccountManagerReconcileRequest",
        "type": "object"
      },
      "AWSAccountManagerTaskResponse": {
        "description": "Response model for POST endpoints.\n\nReturned immediately when task is queued. Contains task_id and status_url\nfor retrieving the result via GET request.",
        "properties": {
          "id": {
            "description": "Task ID",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "default": "pending",
            "description": "Task status (always 'pending' initially)"
          },
          "status_url": {
            "description": "URL to retrieve task result (GET request)",
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status_url"
        ],
        "title": "AWSAccountManagerTaskResponse",
        "type": "object"
      },
...
      "AWSAccountRequest": {
        "description": "Request to create a new AWS account under a payer account.",
        "properties": {
          "email": {
            "description": "Account owner email",
            "title": "Email",
            "type": "string"
          },
          "name": {
            "description": "Account name",
            "title": "Name",
            "type": "string"
          },
          "path": {
            "description": "App-interface path to account request file",
            "title": "Path",
            "type": "string"
          },
          "uid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Existing account ID for takeover scenarios",
            "title": "Uid"
          }
        },
        "required": [
          "name",
          "email",
          "path"
        ],
        "title": "AWSAccountRequest",
        "type": "object"
      },
      "AWSPayerAccount": {
        "description": "A payer (management) account in an AWS organization.",
        "properties": {
          "automation_role": {
            "description": "IAM role ARN for account management",
            "title": "Automation Role",
            "type": "string"
          },
          "automation_token": {
            "$ref": "#/components/schemas/Secret",
            "description": "Secret reference for automation credentials"
          },
          "name": {
            "description": "Account name",
            "title": "Name",
            "type": "string"
          },
          "organization_account_tags": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Default tags for organization accounts",
            "title": "Organization Account Tags",
            "type": "object"
          },
          "resources_default_region": {
            "description": "Default AWS region for API calls",
            "title": "Resources Default Region",
            "type": "string"
          },
          "uid": {
            "description": "AWS account ID",
            "title": "Uid",
            "type": "string"
          }
        },
        "required": [
          "name",
          "uid",
          "automation_token",
          "automation_role",
          "resources_default_region"
        ],
        "title": "AWSPayerAccount",
        "type": "object"
      },
...
      "AccountCreateCompleteAction": {
        "description": "Account creation completed \u2014 contains the created account's name and UID.",
        "properties": {
          "account_name": {
            "description": "Name of the created account",
            "title": "Account Name",
            "type": "string"
          },
          "account_uid": {
            "description": "AWS account ID of the created account",
            "title": "Account Uid",
            "type": "string"
          },
          "action_type": {
            "const": "create_complete",
            "default": "create_complete",
            "title": "Action Type",
            "type": "string"
          },
          "payer_account_name": {
            "description": "Name of the payer account that managed creation",
            "title": "Payer Account Name",
            "type": "string"
          }
        },
        "required": [
          "account_name",
          "payer_account_name",
          "account_uid"
        ],
        "title": "AccountCreateCompleteAction",
        "type": "object"
      },
      "AccountCreateIAMUserAction": {
        "description": "IAM user creation action \u2014 contains account and user details.",
        "properties": {
          "account_name": {
            "description": "Name of the account",
            "title": "Account Name",
            "type": "string"
          },
          "action_type": {
            "const": "create_iam_user",
            "default": "create_iam_user",
            "title": "Action Type",
            "type": "string"
          },
          "detail": {
            "default": "",
            "description": "Additional detail",
            "title": "Detail",
            "type": "string"
          },
          "user_name": {
            "description": "IAM user name created",
            "title": "User Name",
            "type": "string"
          }
        },
        "required": [
          "account_name",
          "user_name"
        ],
        "title": "AccountCreateIAMUserAction",
        "type": "object"
      },
...
      "CreateAccountResult": {
        "description": "Result for GET /create-account/{task_id}.\n\nOn completion, contains the created account's name and UID.",
        "properties": {
          "actions": {
            "description": "Completed account creation (empty while in progress)",
            "items": {
              "$ref": "#/components/schemas/AccountCreateCompleteAction"
            },
            "title": "Actions",
            "type": "array"
          },
          "applied_count": {
            "default": 0,
            "description": "Number of actions actually applied (0 if dry_run=True)",
            "title": "Applied Count",
            "type": "integer"
          },
          "errors": {
            "default": [],
            "description": "List of errors encountered during reconciliation",
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "description": "Task execution status (pending/success/failed)"
          }
        },
        "required": [
          "status"
        ],
        "title": "CreateAccountResult",
        "type": "object"
      },
      "CreateIAMUserResult": {
        "description": "Result for GET /create-iam-user/{task_id}.\n\nContains IAM user creation details.",
        "properties": {
          "actions": {
            "description": "IAM user creation actions performed",
            "items": {
              "$ref": "#/components/schemas/AccountCreateIAMUserAction"
            },
            "title": "Actions",
            "type": "array"
          },
          "applied_count": {
            "default": 0,
            "description": "Number of actions actually applied (0 if dry_run=True)",
            "title": "Applied Count",
            "type": "integer"
          },
          "errors": {
            "default": [],
            "description": "List of errors encountered during reconciliation",
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "description": "Task execution status (pending/success/failed)"
          }
        },
        "required": [
          "status"
        ],
        "title": "CreateIAMUserResult",
        "type": "object"
      },
...
      "ReconcileResult": {
        "description": "Result for GET /reconcile/{task_id}.\n\nSuccess/failure status for account reconciliation.",
        "properties": {
          "actions": {
            "description": "Reconciliation actions performed",
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReconcileActionTag"
                },
                {
                  "$ref": "#/components/schemas/ReconcileActionMoveOU"
                },
                {
                  "$ref": "#/components/schemas/ReconcileActionSetAlias"
                },
                {
                  "$ref": "#/components/schemas/ReconcileActionRequestQuota"
                },
                {
                  "$ref": "#/components/schemas/ReconcileActionEnableSupport"
                },
                {
                  "$ref": "#/components/schemas/ReconcileActionSetSecurityContact"
                },
                {
                  "$ref": "#/components/schemas/ReconcileActionSetRegions"
                }
              ]
            },
            "title": "Actions",
            "type": "array"
          },
          "applied_count": {
            "default": 0,
            "description": "Number of actions actually applied (0 if dry_run=True)",
            "title": "Applied Count",
            "type": "integer"
          },
          "errors": {
            "default": [],
            "description": "List of errors encountered during reconciliation",
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "description": "Task execution status (pending/success/failed)"
          }
        },
        "required": [
          "status"
        ],
        "title": "ReconcileResult",
        "type": "object"
      },
...
      "Secret": {
        "description": "Reference to a secret stored in a secret manager.",
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Specific field within the secret",
            "title": "Field"
          },
          "path": {
            "description": "Path to the secret",
            "title": "Path",
            "type": "string"
          },
          "secret_manager_url": {
            "description": "Secret Manager URL",
            "title": "Secret Manager Url",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Version of the secret",
            "title": "Version"
          }
        },
        "required": [
          "secret_manager_url",
          "path"
        ],
        "title": "Secret",
        "type": "object"
      },
...
      "TaskStatus": {
        "description": "Status for background tasks.\n\nUsed across all async API endpoints to indicate task execution state.",
        "enum": [
          "pending",
          "success",
          "failed"
        ],
        "title": "TaskStatus",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for qontract-reconcile api integrations",
    "summary": "Qontract API OpenAPI schema",
    "title": "Qontract API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/integrations/aws-account-manager/create-account": {
      "post": {
        "description": "Queue creation of a single AWS account.\n\nQueues a background task and returns immediately with a task_id.\nThe task retries itself until the full workflow completes.\nUse GET /create-account/{task_id} to retrieve the result.\n\nReturns 409 with the original task response if a creation workflow\nfor the same request is already in progress.",
        "operationId": "aws-account-manager-create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AWSAccountManagerCreateAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AWSAccountManagerTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AWSAccountManagerTaskResponse"
                }
              }
            },
            "description": "Conflict"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Aws Account Manager Create",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/integrations/aws-account-manager/create-account/{task_id}": {
      "get": {
        "description": "Retrieve account creation task result.",
        "operationId": "aws-account-manager-create-status",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "description": "Optional: Block up to N seconds for completion.",
            "in": "query",
            "name": "timeout",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 300,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional: Block up to N seconds for completion.",
              "title": "Timeout"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccountResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Aws Account Manager Create Status",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/integrations/aws-account-manager/create-iam-user": {
      "post": {
        "description": "Queue creation of an IAM user in an AWS account.\n\nAssumes into the account via the payer's manager role and creates\nan IAM user with the specified policy. Credentials are saved to Vault.\nUse GET /create-iam-user/{task_id} to retrieve the result.",
        "operationId": "aws-account-manager-create-iam-user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AWSAccountManagerCreateIAMUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AWSAccountManagerTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Aws Account Manager Create Iam User",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/integrations/aws-account-manager/create-iam-user/{task_id}": {
      "get": {
        "description": "Retrieve IAM user creation task result.",
        "operationId": "aws-account-manager-create-iam-user-status",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "description": "Optional: Block up to N seconds for completion.",
            "in": "query",
            "name": "timeout",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 300,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional: Block up to N seconds for completion.",
              "title": "Timeout"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateIAMUserResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Aws Account Manager Create Iam User Status",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/integrations/aws-account-manager/reconcile": {
      "post": {
        "description": "Queue reconciliation of a single AWS account.\n\nQueues a background task and returns immediately with a task_id.\nUse GET /reconcile/{task_id} to retrieve the result.",
        "operationId": "aws-account-manager-reconcile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AWSAccountManagerReconcileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AWSAccountManagerTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Aws Account Manager Reconcile",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/v1/integrations/aws-account-manager/reconcile/{task_id}": {
      "get": {
        "description": "Retrieve reconciliation task result.",
        "operationId": "aws-account-manager-reconcile-status",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "description": "Optional: Block up to N seconds for completion.",
            "in": "query",
            "name": "timeout",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 300,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional: Block up to N seconds for completion.",
              "title": "Timeout"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Aws Account Manager Reconcile Status",
        "tags": [
          "integrations"
        ]
      }
    },
...
  }
}